requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
scipy
elasticsearch
nltk
numpy
lda
pandas
sklearn
pymongo
flask
Flask-API
matplotlib
pathos
